updates to multiplayer

+++ FPS1/Assets/Scripts/PlayerWeapon.cs
@@ -2,9 +2,7 @@
 
 [System.Serializable]
 public class PlayerWeapon {
-
     public string name = "Glock";
-
-    public float damage = 10f;
+    public int damage = 10;
     public float range = 100f;
 }
